LRDOCS-5637 Add RSS Publisher to Deprecated Apps
The RSS Publisher widget was deprecated and hidden according to https://issues.liferay.com/browse/LPS-80535.  That ticket also contains a workaround for making the widget visible. The workaround will be documented in a Support k-base article.  I'm not sure how you want to mention the workaround here if at all.